requirements.txt
Django==2.0.5
djangorestframework==3.8.2
pytz==2018.4
gunicorn==19.8.1
django-heroku==0.3.1
django-cors-headers==2.2.0
Pillow==5.2.0